applications/thesis/submitter/requirements.txt
APScheduler==3.10.4
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
Flask==3.0.3
idna==3.7
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
python-decouple==3.8
python-dotenv==1.0.1
python-swiftclient==4.5.0
pytz==2024.1
requests==2.31.0
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0